Move some sumcheck utils to a separate class.
