+++ circuits/cpp/barretenberg/cpp/src/barretenberg/honk/sumcheck/sumcheck_round.hpp
@@ -3,7 +3,9 @@
 #include "barretenberg/common/thread.hpp"
 #include "barretenberg/polynomials/barycentric.hpp"
 #include "barretenberg/polynomials/pow.hpp"
+#include "barretenberg/proof_system/flavor/flavor.hpp"
 #include "barretenberg/proof_system/relations/relation_parameters.hpp"
+#include "barretenberg/proof_system/relations/utils.hpp"
 
 namespace proof_system::honk::sumcheck {
 
@@ -51,6 +53,7 @@ namespace proof_system::honk::sumcheck {
 
 template <typename Flavor> class SumcheckProverRound {
 
+    using Utils = barretenberg::RelationUtils<Flavor>;
     using Relations = typename Flavor::Relations;
     using RelationUnivariates = typename Flavor::RelationUnivariates;
 
@@ -76,7 +79,7 @@ template <typename Flavor> class SumcheckProverRound {
         : round_size(initial_round_size)
     {
         // Initialize univariate accumulators to 0
-        zero_univariates(univariate_accumulators);
+        Utils::zero_univariates(univariate_accumulators);
     }
 
     /**
@@ -89,13 +92,13 @@ template <typename Flavor> class SumcheckProverRound {
         FF challenge, const barretenberg::PowUnivariate<FF>& pow_univariate)
     {
         FF running_challenge = 1;
-        scale_univariates(univariate_accumulators, challenge, running_challenge);
+        Utils::scale_univariates(univariate_accumulators, challenge, running_challenge);
 
         auto result = barretenberg::Univariate<FF, MAX_RANDOM_RELATION_LENGTH>(0);
         extend_and_batch_univariates(univariate_accumulators, pow_univariate, result);
 
         // Reset all univariate accumulators to 0 before beginning accumulation in the next round
-        zero_univariates(univariate_accumulators);
+        Utils::zero_univariates(univariate_accumulators);
         return result;
     }
 
@@ -121,10 +124,11 @@ template <typename Flavor> class SumcheckProverRound {
      * values. Most likely this will end up being S_l(0), ... , S_l(t-1) where t is around 12. At the end, reset all
      * univariate accumulators to be zero.
      */
-    barretenberg::Univariate<FF, MAX_RANDOM_RELATION_LENGTH> compute_univariate(auto& polynomials,
-                                                                  const proof_system::RelationParameters<FF>& relation_parameters,
-                                                                  const barretenberg::PowUnivariate<FF>& pow_univariate,
-                                                                  const FF alpha)
+    barretenberg::Univariate<FF, MAX_RANDOM_RELATION_LENGTH> compute_univariate(
+        auto& polynomials,
+        const proof_system::RelationParameters<FF>& relation_parameters,
+        const barretenberg::PowUnivariate<FF>& pow_univariate,
+        const FF alpha)
     {
         // Precompute the vector of required powers of zeta
         // TODO(luke): Parallelize this
@@ -148,7 +152,7 @@ template <typename Flavor> class SumcheckProverRound {
         // Constuct univariate accumulator containers; one per thread
         std::vector<RelationUnivariates> thread_univariate_accumulators(num_threads);
         for (auto& accum : thread_univariate_accumulators) {
-            zero_univariates(accum);
+            Utils::zero_univariates(accum);
         }
 
         // Constuct extended edge containers; one per thread
@@ -180,7 +184,7 @@ template <typename Flavor> class SumcheckProverRound {
 
         // Accumulate the per-thread univariate accumulators into a single set of accumulators
         for (auto& accumulators : thread_univariate_accumulators) {
-            add_nested_tuples(univariate_accumulators, accumulators);
+            Utils::add_nested_tuples(univariate_accumulators, accumulators);
         }
         // Batch the univariate contributions from each sub-relation to obtain the round univariate
         return batch_over_relations(alpha, pow_univariate);
@@ -262,105 +266,7 @@ template <typename Flavor> class SumcheckProverRound {
                 result += extended;
             }
         };
-        apply_to_tuple_of_tuples(tuple, extend_and_sum);
-    }
-
-    /**
-     * @brief Set all coefficients of Univariates to zero
-     *
-     * @details After computing the round univariate, it is necessary to zero-out the accumulators used to compute it.
-     */
-    static void zero_univariates(auto& tuple)
-    {
-        auto set_to_zero = []<size_t, size_t>(auto& element) {
-            std::fill(element.evaluations.begin(), element.evaluations.end(), FF(0));
-        };
-        apply_to_tuple_of_tuples(tuple, set_to_zero);
-    }
-
-    /**
-     * @brief Scale Univaraites by consecutive powers of the provided challenge
-     *
-     * @param tuple Tuple of tuples of Univariates
-     * @param challenge
-     * @param current_scalar power of the challenge
-     */
-    static void scale_univariates(auto& tuple, const FF& challenge, FF current_scalar)
-    {
-        auto scale_by_consecutive_powers_of_challenge = [&]<size_t, size_t>(auto& element) {
-            element *= current_scalar;
-            current_scalar *= challenge;
-        };
-        apply_to_tuple_of_tuples(tuple, scale_by_consecutive_powers_of_challenge);
-    }
-
-    /**
-     * @brief General purpose method for applying an operation to a tuple of tuples of Univariates
-     *
-     * @tparam Operation Any operation valid on Univariates
-     * @tparam outer_idx Index into the outer tuple
-     * @tparam inner_idx Index into the inner tuple
-     * @param tuple A Tuple of tuples of Univariates
-     * @param operation Operation to apply to Univariates
-     */
-    template <class Operation, size_t outer_idx = 0, size_t inner_idx = 0>
-    static void apply_to_tuple_of_tuples(auto& tuple, Operation&& operation)
-    {
-        auto& inner_tuple = std::get<outer_idx>(tuple);
-        auto& univariate = std::get<inner_idx>(inner_tuple);
-
-        // Apply the specified operation to each Univariate
-        operation.template operator()<outer_idx, inner_idx>(univariate);
-
-        const size_t inner_size = std::tuple_size_v<std::decay_t<decltype(std::get<outer_idx>(tuple))>>;
-        const size_t outer_size = std::tuple_size_v<std::decay_t<decltype(tuple)>>;
-
-        // Recurse over inner and outer tuples
-        if constexpr (inner_idx + 1 < inner_size) {
-            apply_to_tuple_of_tuples<Operation, outer_idx, inner_idx + 1>(tuple, std::forward<Operation>(operation));
-        } else if constexpr (outer_idx + 1 < outer_size) {
-            apply_to_tuple_of_tuples<Operation, outer_idx + 1, 0>(tuple, std::forward<Operation>(operation));
-        }
-    }
-
-    /**
-     * @brief Componentwise addition of two tuples
-     * @details Used for adding tuples of Univariates but in general works for any object for which += is
-     * defined. The result is stored in the first tuple.
-     *
-     * @tparam T Type of the elements contained in the tuples
-     * @param tuple_1 First summand. Result stored in this tuple
-     * @param tuple_2 Second summand
-     */
-    template <typename... T>
-    static constexpr void add_tuples(std::tuple<T...>& tuple_1, const std::tuple<T...>& tuple_2)
-    {
-        auto add_tuples_helper = [&]<std::size_t... I>(std::index_sequence<I...>)
-        {
-            ((std::get<I>(tuple_1) += std::get<I>(tuple_2)), ...);
-        };
-
-        add_tuples_helper(std::make_index_sequence<sizeof...(T)>{});
-    }
-
-    /**
-     * @brief Componentwise addition of nested tuples (tuples of tuples)
-     * @details Used for summing tuples of tuples of Univariates. Needed for Sumcheck multithreading. Each thread
-     * accumulates realtion contributions across a portion of the hypecube and then the results are accumulated into a
-     * single nested tuple.
-     *
-     * @tparam Tuple
-     * @tparam Index Index into outer tuple
-     * @param tuple_1 First nested tuple summand. Result stored here
-     * @param tuple_2 Second summand
-     */
-    template <typename Tuple, std::size_t Index = 0>
-    static constexpr void add_nested_tuples(Tuple& tuple_1, const Tuple& tuple_2)
-    {
-        if constexpr (Index < std::tuple_size<Tuple>::value) {
-            add_tuples(std::get<Index>(tuple_1), std::get<Index>(tuple_2));
-            add_nested_tuples<Tuple, Index + 1>(tuple_1, tuple_2);
-        }
+        Utils::apply_to_tuple_of_tuples(tuple, extend_and_sum);
     }
 };
 
@@ -440,12 +346,12 @@ template <typename Flavor> class SumcheckVerifierRound {
      * @param round_challenge u_l
      * @return FF sigma_{l+1} = S^l(u_l)
      */
-    FF compute_next_target_sum(barretenberg::Univariate<FF, MAX_RANDOM_RELATION_LENGTH>& univariate, FF& round_challenge)
+    FF compute_next_target_sum(barretenberg::Univariate<FF, MAX_RANDOM_RELATION_LENGTH>& univariate,
+                               FF& round_challenge)
     {
         // IMPROVEMENT(Cody): Use barycentric static method, maybe implement evaluation as member
         // function on Univariate.
-        auto barycentric =
-            barretenberg::BarycentricData<FF, MAX_RANDOM_RELATION_LENGTH, MAX_RANDOM_RELATION_LENGTH>();
+        auto barycentric = barretenberg::BarycentricData<FF, MAX_RANDOM_RELATION_LENGTH, MAX_RANDOM_RELATION_LENGTH>();
         // Evaluate T^{l}(u_{l})
         target_total_sum = barycentric.evaluate(univariate, round_challenge);
