+++ circuits/cpp/barretenberg/cpp/src/barretenberg/proof_system/relations/utils.hpp
@@ -1,0 +1,135 @@
+#pragma once
+#include "barretenberg/common/thread.hpp"
+#include "barretenberg/ecc/curves/bn254/fr.hpp"
+#include "barretenberg/polynomials/barycentric.hpp"
+#include "barretenberg/polynomials/pow.hpp"
+#include "barretenberg/proof_system/relations/relation_parameters.hpp"
+
+namespace barretenberg {
+
+template <typename Flavor> class RelationUtils {
+  public:
+    using FF = typename Flavor::FF;
+    using Relations = typename Flavor::Relations;
+    using RelationUnivariates = typename Flavor::RelationUnivariates;
+
+    /**
+     * @brief Set all coefficients of Univariates to zero
+     *
+     * @details After computing the round univariate, it is necessary to zero-out the accumulators used to compute it.
+     * WORKTODO: rename
+     */
+    static void zero_univariates(auto& tuple)
+    {
+        auto set_to_zero = []<size_t, size_t>(auto& element) {
+            std::fill(element.evaluations.begin(), element.evaluations.end(), FF(0));
+        };
+        apply_to_tuple_of_tuples(tuple, set_to_zero);
+    }
+
+    /**
+     * @brief Scale Univaraites by consecutive powers of the provided challenge
+     *
+     * @param tuple Tuple of tuples of Univariates
+     * @param challenge
+     * @param current_scalar power of the challenge
+     */
+    static void scale_univariates(auto& tuple, const FF& challenge, FF current_scalar)
+    {
+        auto scale_by_consecutive_powers_of_challenge = [&]<size_t, size_t>(auto& element) {
+            element *= current_scalar;
+            current_scalar *= challenge;
+        };
+        apply_to_tuple_of_tuples(tuple, scale_by_consecutive_powers_of_challenge);
+    }
+
+    /**
+     * @brief General purpose method for applying an operation to a tuple of tuples of Univariates
+     *
+     * @tparam Operation Any operation valid on Univariates
+     * @tparam outer_idx Index into the outer tuple
+     * @tparam inner_idx Index into the inner tuple
+     * @param tuple A Tuple of tuples of Univariates
+     * @param operation Operation to apply to Univariates
+     */
+    template <class Operation, size_t outer_idx = 0, size_t inner_idx = 0>
+    static void apply_to_tuple_of_tuples(auto& tuple, Operation&& operation)
+    {
+        auto& inner_tuple = std::get<outer_idx>(tuple);
+        auto& univariate = std::get<inner_idx>(inner_tuple);
+
+        // Apply the specified operation to each Univariate
+        operation.template operator()<outer_idx, inner_idx>(univariate);
+
+        const size_t inner_size = std::tuple_size_v<std::decay_t<decltype(std::get<outer_idx>(tuple))>>;
+        const size_t outer_size = std::tuple_size_v<std::decay_t<decltype(tuple)>>;
+
+        // Recurse over inner and outer tuples
+        if constexpr (inner_idx + 1 < inner_size) {
+            apply_to_tuple_of_tuples<Operation, outer_idx, inner_idx + 1>(tuple, std::forward<Operation>(operation));
+        } else if constexpr (outer_idx + 1 < outer_size) {
+            apply_to_tuple_of_tuples<Operation, outer_idx + 1, 0>(tuple, std::forward<Operation>(operation));
+        }
+    }
+
+    /**
+     * @brief Componentwise addition of two tuples
+     * @details Used for adding tuples of Univariates but in general works for any object for which += is
+     * defined. The result is stored in the first tuple.
+     *
+     * @tparam T Type of the elements contained in the tuples
+     * @param tuple_1 First summand. Result stored in this tuple
+     * @param tuple_2 Second summand
+     */
+    template <typename... T>
+    static constexpr void add_tuples(std::tuple<T...>& tuple_1, const std::tuple<T...>& tuple_2)
+    {
+        auto add_tuples_helper = [&]<std::size_t... I>(std::index_sequence<I...>)
+        {
+            ((std::get<I>(tuple_1) += std::get<I>(tuple_2)), ...);
+        };
+
+        add_tuples_helper(std::make_index_sequence<sizeof...(T)>{});
+    }
+
+    /**
+     * @brief Componentwise addition of nested tuples (tuples of tuples)
+     * @details Used for summing tuples of tuples of Univariates. Needed for Sumcheck multithreading. Each thread
+     * accumulates realtion contributions across a portion of the hypecube and then the results are accumulated into a
+     * single nested tuple.
+     *
+     * @tparam Tuple
+     * @tparam Index Index into outer tuple
+     * @param tuple_1 First nested tuple summand. Result stored here
+     * @param tuple_2 Second summand
+     */
+    template <typename Tuple, std::size_t Index = 0>
+    static constexpr void add_nested_tuples(Tuple& tuple_1, const Tuple& tuple_2)
+    {
+        if constexpr (Index < std::tuple_size<Tuple>::value) {
+            add_tuples(std::get<Index>(tuple_1), std::get<Index>(tuple_2));
+            add_nested_tuples<Tuple, Index + 1>(tuple_1, tuple_2);
+        }
+    }
+
+    // /**
+    //  * @brief Given a tuple t = (t_0, t_1, ..., t_{NUM_RELATIONS-1}) and a challenge α,
+    //  * return t_0 + αt_1 + ... + α^{NUM_RELATIONS-1}t_{NUM_RELATIONS-1}).
+    //  *
+    //  * @tparam T : In practice, this is a Univariate<FF, MAX_NUM_RELATIONS>.
+    //  */
+    // template <typename T> T batch_over_relations(auto& univariate_accumulators, FF challenge, const
+    // barretenberg::PowUnivariate<FF>& pow_univariate)
+    // {
+    //     FF running_challenge = 1;
+    //     scale_univariates(univariate_accumulators, challenge, running_challenge);
+
+    //     auto result = barretenberg::Univariate<FF, MAX_RANDOM_RELATION_LENGTH>(0);
+    //     extend_and_batch_univariates(univariate_accumulators, pow_univariate, result);
+
+    //     // Reset all univariate accumulators to 0 before beginning accumulation in the next round
+    //     zero_univariates(univariate_accumulators);
+    //     return result;
+    // }
+};
+} // namespace barretenberg
